Split logging module and test initialization (fixes #78)
