+++ cliquet/__init__.py
@@ -1,7 +1,7 @@
 """Main entry point
 """
-import datetime
 import json
+import datetime
 
 from dateutil import parser as dateparser
 import pkg_resources
@@ -13,11 +13,9 @@ from pyramid.security import NO_PERMISSION_REQUIRED
 
 from cliquet import authentication
 from cliquet import errors
-from cliquet import utils
+from cliquet import logging as cliquet_logging
 from cliquet.session import SessionCache
-from cliquet.utils import msec_time
 
-import structlog
 from cornice import Service
 
 # Monkey Patch Cornice Service to setup the global CORS configuration.
@@ -32,9 +30,6 @@ __version__ = pkg_resources.get_distribution(__package__).version
 
 # The API version is derivated from the module version.
 API_VERSION = 'v%s' % __version__.split('.')[0]
-
-# Main cliquet logger
-logger = structlog.get_logger(__name__)
 
 
 def handle_api_redirection(config):
@@ -141,64 +136,11 @@ def end_of_life_tween_factory(handler, registry):
     return eos_tween
 
 
-def setup_logging(config):
-    """Setup structured logging, and emit `request.summary` event on each
-    request, as recommanded by Mozilla Services standard:
-
-    * https://mana.mozilla.org/wiki/display/CLOUDSERVICES/Logging+Standard
-    * http://12factor.net/logs
-    """
-    # XXX: if debug, render as key=value
-    structlog.configure(
-        processors=[
-            # XXX https://github.com/hynek/structlog/pull/50
-            # structlog.stdlib.filter_by_level,
-            structlog.processors.format_exc_info,
-            utils.MozillaHekaRenderer(config.get_settings())
-        ],
-        # Share the logger context by thread
-        context_class=structlog.threadlocal.wrap_dict(dict),
-        # Integrate with Pyramid logging facilities
-        logger_factory=structlog.stdlib.LoggerFactory())
-
-    def on_new_request(event):
-        request = event.request
-        # Save the time the request was received by the server.
-        event.request._received_at = msec_time()
-
-        # New logger context, with infos for request summary logger.
-        logger.new(agent=request.headers.get('User-Agent'),
-                   path=event.request.path,
-                   method=request.method,
-                   uid=request.authenticated_userid,
-                   lang=request.headers.get('Accept-Language'),
-                   errno=None)
-
-    config.add_subscriber(on_new_request, NewRequest)
-
-    def on_new_response(event):
-        response = event.response
-        request = event.request
-
-        # Compute the request processing time in msec (-1 if unknown)
-        current = msec_time()
-        duration = current - getattr(request, '_received_at', current - 1)
-
-        # Bind infos for request summary logger.
-        logger.bind(code=response.status_code,
-                    t=duration)
-
-        # Ouput application request summary.
-        logger.info('request.summary')
-
-    config.add_subscriber(on_new_response, NewResponse)
-
-
 def includeme(config):
     settings = config.get_settings()
 
-    # Configure logging globally.
-    setup_logging(config)
+    # Configure cliquet logging.
+    cliquet_logging.setup_logging(config)
 
     handle_api_redirection(config)
     config.add_tween("cliquet.end_of_life_tween_factory")
